Move components outside pages
