+++ src/components/SearchBox.tsx
@@ -2,7 +2,7 @@ import { InputAdornment, TextField } from "@material-ui/core";
 import { SearchIcon } from "@saleor/macaw-ui";
 import { ChangeEvent } from "react";
 import { useSearchBox } from "react-instantsearch-hooks-web";
-import styles from "./search.module.css";
+import styles from "../styles/search.module.css";
 
 export function SearchBox() {
     const { query, refine } = useSearchBox();
